refactor gtests and add make test
