+++ ptarm/tests/_testinc_template.cpp
@@ -5,11 +5,12 @@
 
 ////////////////////////////////////////////////////////////////////////
 
-class keys_native: public testing::Test {
+class xxx: public testing::Test {
 protected:
     virtual void SetUp() {
         //RESET_FAKE(external_function)
-        ptarm_init(PTARM_TESTNET, true);
+        ptarm_dbg_malloc_cnt_reset();
+        ptarm_init(PTARM_TESTNET, false);
     }
 
     virtual void TearDown() {
@@ -28,3 +29,9 @@ public:
 };
 
 ////////////////////////////////////////////////////////////////////////
+
+TEST_F(xxx, init)
+{
+    ASSERT_EQ(0, 0);
+    ASSERT_TRUE(NULL != buf.buf);
+}
