+++ ptarm/tests/testinc_hash.cpp
