+++ ptarm/tests/testinc_ln.cpp
@@ -5,11 +5,12 @@
 
 ////////////////////////////////////////////////////////////////////////
 
-class xxx: public testing::Test {
+class ln: public testing::Test {
 protected:
     virtual void SetUp() {
         //RESET_FAKE(external_function)
-        ptarm_init(PTARM_TESTNET, false);
+        ptarm_dbg_malloc_cnt_reset();
+        ptarm_init(PTARM_TESTNET, true);
     }
 
     virtual void TearDown() {
@@ -29,8 +30,12 @@ public:
 
 ////////////////////////////////////////////////////////////////////////
 
-TEST_F(xxx, init)
-{
-    ASSERT_EQ(0, 0);
-    ASSERT_TRUE(NULL != buf.buf);
-}
+//TEST_F(ln, init)
+//{
+//    ptarm_ln_init();
+//}
+//
+//TEST_F(ln, term)
+//{
+//    ptarm_ln_term();
+//}
