+++ ptarm/tests/testinc_ln_bolt4.cpp
@@ -17,6 +17,7 @@ class onion: public testing::Test {
 protected:
     virtual void SetUp() {
         //RESET_FAKE(external_function)
+        ptarm_dbg_malloc_cnt_reset();
         ptarm_init(PTARM_TESTNET, true);
         ptarm_buf_init(&sOnionBuffer);
         spEphPubkey = NULL;
