+++ ptarm/tests/testinc_ln_bolt8.cpp
@@ -9,6 +9,7 @@ class bolt8test: public testing::Test {
 protected:
     virtual void SetUp() {
         //RESET_FAKE(external_function)
+        ptarm_dbg_malloc_cnt_reset();
         ptarm_init(PTARM_TESTNET, true);
     }
